+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### 1.3.4 (2025-09-17)
 1. fix `morebuiltins.utils.gen_id` wrong length issue.
 2. fix `morebuiltins.funcs.LineProfiler` to handle nested function calls correctly.
+3. update `morebuiltins.cmd.log_server` to support close idle file handlers after 300 seconds.
+4. add `get_logger` to `morebuiltins.cmd.log_server` to get a logger with `SocketHandler` to send logs to `LogServer`.
 
 ### 1.3.3 (2025-09-01)
 1. add `morebuiltins.logs.LogHelper.handle_crash` to log uncaught exceptions.